requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.2.0
    # via streamlit
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
attrs==23.2.0
    # via
    #   jsonschema
    #   referencing
blinker==1.7.0
    # via streamlit
cachetools==5.3.3
    # via streamlit
certifi==2024.2.2
    # via
    #   elastic-transport
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via streamlit
colorama==0.4.6
    # via
    #   click
    #   tqdm
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
distro==1.9.0
    # via openai
elastic-transport==8.12.0
    # via elasticsearch
elasticsearch==8.12.1
    # via -r requirements.in
filelock==3.13.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.49.0
    # via matplotlib
fsspec==2024.2.0
    # via
    #   huggingface-hub
    #   torch
gitdb==4.0.11
    # via gitpython
gitpython==3.1.42
    # via streamlit
h11==0.14.0
    # via httpcore
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via openai
huggingface-hub==0.21.3
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==7.0.1
    # via streamlit
iterative-stratification==0.1.7
    # via -r requirements.in
jinja2==3.1.3
    # via
    #   altair
    #   pydeck
    #   torch
joblib==1.3.2
    # via scikit-learn
jsonschema==4.21.1
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
kiwisolver==1.4.5
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
matplotlib==3.8.3
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via torch
numpy==1.26.4
    # via
    #   -r requirements.in
    #   altair
    #   contourpy
    #   iterative-stratification
    #   matplotlib
    #   pandas
    #   pyarrow
    #   pydeck
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   streamlit
    #   transformers
openai==1.13.3
    # via -r requirements.in
packaging==23.2
    # via
    #   altair
    #   huggingface-hub
    #   matplotlib
    #   streamlit
    #   transformers
pandas==2.2.1
    # via
    #   -r requirements.in
    #   altair
    #   streamlit
pillow==10.2.0
    # via
    #   matplotlib
    #   sentence-transformers
    #   streamlit
protobuf==4.25.3
    # via streamlit
pyarrow==15.0.0
    # via streamlit
pydantic==2.6.3
    # via openai
pydantic-core==2.16.3
    # via pydantic
pydeck==0.8.1b0
    # via streamlit
pygments==2.17.2
    # via rich
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
    #   streamlit
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   transformers
referencing==0.33.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2023.12.25
    # via transformers
requests==2.31.0
    # via
    #   huggingface-hub
    #   streamlit
    #   transformers
rich==13.7.0
    # via streamlit
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
safetensors==0.4.2
    # via transformers
scikit-learn==1.4.1.post1
    # via
    #   -r requirements.in
    #   iterative-stratification
    #   sentence-transformers
scipy==1.12.0
    # via
    #   -r requirements.in
    #   iterative-stratification
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==2.5.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
streamlit==1.31.1
    # via -r requirements.in
sympy==1.12
    # via torch
tenacity==8.2.3
    # via streamlit
threadpoolctl==3.3.0
    # via scikit-learn
tokenizers==0.15.2
    # via transformers
toml==0.10.2
    # via streamlit
toolz==0.12.1
    # via altair
torch==2.2.1
    # via sentence-transformers
tornado==6.4
    # via streamlit
tqdm==4.66.2
    # via
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.38.2
    # via sentence-transformers
typing==3.7.4.3
    # via -r requirements.in
typing-extensions==4.10.0
    # via
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   streamlit
    #   torch
tzdata==2024.1
    # via
    #   pandas
    #   tzlocal
tzlocal==5.2
    # via streamlit
urllib3==2.2.1
    # via
    #   elastic-transport
    #   requests
validators==0.22.0
    # via streamlit
watchdog==4.0.0
    # via streamlit
zipp==3.17.0
    # via importlib-metadata
